Moving build system to maven
Added a build.txt to help explain usage.

The tests, antlr and build stuff has been moved into the standard mvn structure (src/...), but
the rest of the source is still in java/src. For right now the build system is fine with it, but it should
also be moved at some point soon. Also, removed all the .project/.settings/.classpath files since
maven import into eclipse makes that work correctly without messing with that stuff.

I'm leaving the build/ directory for the moment, until we are sure mvn works, but it should also
go away soon (when we are confident).

Additionally, adding .gitignore that's correct for maven