Adding hbase-stats project to contrib/
This is the basis for supporting equal-depth histograms per region, guideposts, #49.
Its basically a 0.94 backport of the code attached to HBASE-7958. Logically, there is
no real difference, though the implementation is has some slight changes as on trunk there
is no need for a coprocessor as the stats gather is all built in.

Currently, HBASE-7958 is stil open for review and depends on the the system tables (HBASE-7999) or
namespaces (HBASE-8105) patches. We resolve this by just giving the stats table a special name
that should be fairly close to the pending name for the stats table in HBASE-7958. In the event
that we cannot maintain the same table name for the stats table, we do have an opportunity to
copy over the data to the neew table as there is currently a required downtime to upgrade from 0.94
to 0.96.

This is actually a bit more advanced than the posted patch - a lot more work has gone into usability
and verifying correctness. Further, there are some obvious changes as we need to support coprocessors
rather than built options (but that is actually a relatively minor change).