Lazy-load selected columns for current row state
Before, we were loading all the columns of the row into memory.
This could be wildly inefficient as we just care about a small
subset of the columns (usually) for indexing. Now, we just load the
columns the user requests per-index, on demand.

This lead to an issue (esp when combined with the local HRegion reads)
where two KVs with the same row-key (r,f,q,ts) would sort in the wrong
order based on memstore timestamp. Resolved this with a custom KV
comparator that always sorts the pending updates (as PendingKeyValue)
before regular keyvalues, in the case that everything is the same.

Really, we should probably also rip out the ExposedMemStore and go
with a cleaner, simpler implementation. That's a V2 thing, but adding
comments to this effect.