Add closing to scanners.
Scanners were not being closed, causing a resource leak on the
underlying MemStore. While this was fine for short operations
(the ExposedMemStore was the only reference and it was let go after
building the index updates for each mutation), it could be an issue
for longer operations or operations that need to create a lot of
scanners (e.g. lots of indexes).

Now the scanners are allowed to be GC'ed.