Speeding up IndexMemStore, removing ExposedMemStore.
Lots of time was spent creating a config in IndexMemStore that we didn't need - removing it.
Also, removing the timetracker stuff - we don't care about whether or not to get a scanner
for the store, given a timerange - we always scan the store.

Additionally, removing ExposedMemStore since we are using our own impl now.