Late-replaying index writes from the WAL.
When WAL recovery occurs, we need to ensure that we can write
to the index table. If the index region we were targeting lived
on the same server as the primary region, it will likely be involved
in the same log splitting and recovery process. This means, when we
replay the writes into the primary table, the index table/region
we would write to is probably not up yet. In 0.96 this is fixed by
allowing regions to take writes before reads during WAL replay. We
are only on 0.94 (and its a huge effort to backport that feature),
so instead we can get around the issue by buffering the failed writes
and replaying them later (in the postOpen hook).