Fixing (un)compressed WAL Tests.
Major issue was conflict between the IndexWALReader (which supports uncompressed WAL) and IndexWALEditCodec
which supports compressed and uncompressed kvs. When the reader is installed, it can only support uncompressed
WALEdits, which means using the Codec is unnecessary (and vice-versa). Therefore, in the current version, we
only want to support the IndexWALEditCodec. However, we keed around the IndexWALReader to support 0.94.9, for
those who want it (and updating the test to indicate that).