Adding paralell builder to concurrently run index builds + threading fixes.
Moving building to an IndexBuildManager. This helps abstract some of the details of
building away from the Indexer. Also, means we can leak less of the details of the
implementation away (though not as many as I'd like)

Adding some more logging and interruptibility to acquiring the INDEX_UPDATE lock
to help minimize locking risks for server shutdown.