Abstracting TaskRunner logic.
Helps for testing and also later adding other TaskRunners

Additionally, formalizing Task/TaskBatch to support better failure policies.

This way we can tie all the elements of the batch together and let the tasks determine
if they want to fail together if another task fails in the same batch. We then also
update the batch failure when we find a single task failure in the QuickFailing
implementation.

Additionally, this supports more complex task runners, for instance the task runner
that would be needed for the TrackingParallelWriteIndexCommitter (not yet implemented
as it would just be too much in the same patch