Cleanup up defaults + logging.
This resolves the issue of salted indexes. Essentially, HTables have default pools with unbounded
threads, but direct handoff of tasks via the queue. Therefore, the default of 1 thread meant
that batching updates in a multi-threaded environment (or even to multiple RS) would cause
a RejectedExecutionException. We up the value to Integer.MAX_VALUE by default, allowing us
to grow as we need.

This is not an ideal solution - it can quickly burst up to a large number of threads. What we
really want to do is set a max level of parallelization and then queue up any remaining tasks
(similar to what we are doing with the rest of the indexing pools).

Additionally, we support the standard HTable retry values in a "set if not set" style so
the user can override their settings. This means you have increased flexibility in how quickly
you want an index update to fail, making it more resilient to RS (particularly the one holding
-ROOT- and/or .META.) outage.